cloud_function/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
anyio==4.9.0
    # via httpx
blinker==1.9.0
    # via flask
cachecontrol==0.14.3
    # via firebase-admin
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   flask
    #   functions-framework
cloudevents==1.12.0
    # via functions-framework
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==45.0.5
    # via pyjwt
deprecation==2.1.0
    # via cloudevents
firebase-admin==7.0.0
    # via homelab-cloud-controller
flask==3.0.3
    # via
    #   functions-framework
    #   homelab-cloud-controller
functions-framework==3.8.2
    # via homelab-cloud-controller
google-api-core==2.25.1
    # via
    #   firebase-admin
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-error-reporting
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-error-reporting
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-storage
    #   homelab-cloud-controller
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-error-reporting==1.11.1
    # via homelab-cloud-controller
google-cloud-firestore==2.21.0 ; platform_python_implementation != 'PyPy'
    # via firebase-admin
google-cloud-logging==3.12.1
    # via google-cloud-error-reporting
google-cloud-pubsub==2.31.0
    # via homelab-cloud-controller
google-cloud-storage==3.2.0
    # via firebase-admin
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-logging
    #   google-cloud-pubsub
grpcio==1.74.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.2
    # via
    #   google-api-core
    #   google-cloud-pubsub
gunicorn==23.0.0 ; sys_platform != 'win32'
    # via functions-framework
h11==0.16.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via firebase-admin
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==2.1.3
    # via
    #   homelab-cloud-controller
    #   jinja2
    #   werkzeug
msgpack==1.1.1
    # via cachecontrol
opentelemetry-api==1.35.0
    # via
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.35.0
    # via google-cloud-pubsub
opentelemetry-semantic-conventions==0.56b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   deprecation
    #   gunicorn
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-error-reporting
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-pubsub
protobuf==5.29.5
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-error-reporting
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22 ; platform_python_implementation != 'PyPy'
    # via cffi
pyjwt==2.10.1
    # via firebase-admin
requests==2.32.4
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
typing-extensions==4.14.1
    # via
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.3
    # via
    #   flask
    #   functions-framework
zipp==3.23.0
    # via importlib-metadata
